% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nadarayaWatson-package.R
\docType{package}
\name{nadarayaWatson-package}
\alias{nadarayaWatson}
\alias{nadarayaWatson-package}
\title{nadarayaWatson: Rcpp Based Nadaraya-Watson Kernel Regression}
\description{
This package uses Rcpp to implement the Nadaraya-Watson kernel regression. The Nadaraya-Watson kernel regression is a non-parametric regression method that estimates the conditional expectation of a random variable given another random variable.
}
\author{
\strong{Maintainer}: Xuyuan Zhang \email{zxuyuan@umich.edu}

}
\keyword{internal}
